+++ app/src/main/java/com/bitchat/android/model/BitchatMessage.kt
@@ -342,71 +342,4 @@ data class BitchatMessage(
     }
 }
 
-/**
- * Delivery acknowledgment structure - exact same as iOS version
- */
-@Parcelize
-data class DeliveryAck(
-    val originalMessageID: String,
-    val ackID: String = UUID.randomUUID().toString(),
-    val recipientID: String,
-    val recipientNickname: String,
-    val timestamp: Date = Date(),
-    val hopCount: UInt
-) : Parcelable {
 
-    private val gson = GsonBuilder()
-        .setDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'")
-        .create()
-    
-    fun encode(): ByteArray? {
-        return try {
-            gson.toJson(this).toByteArray(Charsets.UTF_8)
-        } catch (e: Exception) {
-            null
-        }
-    }
-    
-    companion object {
-        fun decode(data: ByteArray): DeliveryAck? {
-            return try {
-                val json = String(data, Charsets.UTF_8)
-                com.google.gson.Gson().fromJson(json, DeliveryAck::class.java)
-            } catch (e: Exception) {
-                null
-            }
-        }
-    }
-}
-
-/**
- * Read receipt structure - exact same as iOS version
- */
-@Parcelize
-data class ReadReceipt(
-    val originalMessageID: String,
-    val receiptID: String = UUID.randomUUID().toString(),
-    val readerID: String,
-    val readerNickname: String,
-    val timestamp: Date = Date()
-) : Parcelable {
-    
-    fun encode(): ByteArray? {
-        return try {
-            com.google.gson.Gson().toJson(this).toByteArray(Charsets.UTF_8)
-        } catch (e: Exception) {
-            null
-        }
-    }
-    
-    companion object {
-        fun decode(data: ByteArray): ReadReceipt? {
-            return try {
-                val json = String(data, Charsets.UTF_8)
-                com.google.gson.Gson().fromJson(json, ReadReceipt::class.java)
-            } catch (e: Exception) {
-                null
-            }
-        }
-    }
-}
